Clean code

+++ src/Controller/Game/GameController.php
@@ -11,7 +11,6 @@ use App\Entity\GamePublishers;
 use App\Entity\GameTheme;
 use App\Entity\GameSerie;
 use App\Form\GameAddType;
-use App\Form\GameType;
 use App\Repository\GameDeveloperRepository;
 use App\Repository\GameGenreRepository;
 use App\Repository\GamePlatformRepository;
@@ -20,11 +19,13 @@ use App\Repository\GameRepository;
 use App\Repository\GameModeRepository;
 use App\Repository\GameSerieRepository;
 use App\Repository\GameThemeRepository;
+use App\Service\FileService;
+use App\Service\IGDBService;
 use App\Service\StrSpecialCharsLower;
 use App\Service\TimeService;
 use DateTime;
 use Doctrine\ORM\EntityManagerInterface;
-use GuzzleHttp\Client;
+use Exception;
 use GuzzleHttp\Exception\GuzzleException;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -46,10 +47,15 @@ class GameController extends AbstractController
 
 
     #[Route('/details/{id}', name: 'game_details', methods: ['GET'])]
-    public function show(GameRepository $gameRepository, TimeService $timeService, $id): Response
+    public function show(GameRepository $gameRepository, int $id): Response
     {
 
         $game = $gameRepository->find($id);
+
+        if (!$game){
+            $this->addFlash('error', 'Ce jeu n\'existe pas');
+            return $this->redirectToRoute('game');
+        }
 
         return $this->render('game/game/details.html.twig', [
             'game' => $game,
@@ -84,22 +90,26 @@ class GameController extends AbstractController
 
 
     #[Route('/delete/{id}', name: 'game_delete')]
-    public function delete(Request $request, EntityManagerInterface $entityManager, GameRepository $gameRepository, $id): Response
+    public function delete(EntityManagerInterface $entityManager, GameRepository $gameRepository, int $id): Response
     {
 
         $game = $gameRepository->find($id);
 
-        $gameTrackers = $game->getGameTrackers();
+        if($game){
 
-        foreach ($gameTrackers as $gameTracker){
+            $gameTrackers = $game->getGameTrackers();
 
-            $entityManager->remove($gameTracker);
+            foreach ($gameTrackers as $gameTracker) {
+
+                $entityManager->remove($gameTracker);
+
+            }
+
+            $entityManager->remove($game);
+
+            $entityManager->flush();
 
         }
-
-        $entityManager->remove($game);
-
-        $entityManager->flush();
 
         return $this->redirectToRoute('game', [], Response::HTTP_SEE_OTHER);
     }
@@ -120,6 +130,8 @@ class GameController extends AbstractController
         GamePlatformRepository   $gamePlatformRepository,
         GameSerieRepository      $gameSerieRepository,
         StrSpecialCharsLower     $strSpecialCharsLower,
+        FileService              $fileService,
+        IGDBService              $IGDBService,
         KernelInterface          $kernel
     ): Response
     {
@@ -134,56 +146,27 @@ class GameController extends AbstractController
             $platformId = $formData['platforms']->getIgdbID();
             $idGame = $formData['igdbId'];
 
-            // AUTHENTIFICATION
-
-            $client = new Client();
-
-            $response = $client->post("https://id.twitch.tv/oauth2/token?client_id=sd5xdt5w2lkjr7ws92fxjdlicvb5u2&client_secret=tymefepntjuva1n9ipa3lkjts2pmdh&grant_type=client_credentials", [
-                'headers' => [
-                    'Content-Type' => 'application/json',
-                    'Accept' => 'application/json',
-                ]
-            ]);
-
-            $data = json_decode($response->getBody(), true);
-
-            $token = "Bearer ".$data['access_token'];
-
             $game = new Game();
 
             // NAME
 
-            $response = $client->post("https://api.igdb.com/v4/games", [
-                'headers' => [
-                    'Content-Type' => 'application/json',
-                    'Accept' => 'application/json',
-                    'Client-ID' => 'sd5xdt5w2lkjr7ws92fxjdlicvb5u2',
-                    'Authorization' => $token
-                ],
-                'body' => 'fields name,game_modes,genres,themes; where id = '.$idGame.';'
-            ]);
+            $body = 'fields name,game_modes,genres,themes; where id = '.$idGame.';';
 
-            $dataGame = json_decode($response->getBody(), true)[0];
+            $dataGame = $IGDBService->getData('games', $body);
 
             $game->setName($dataGame['name']);
 
             $idParent = null;
 
-            if (isset($data['version_parent'])) {
-                $idParent = $data['version_parent'];
+            if (isset($dataGame['version_parent'])) {
+                $idParent = $dataGame['version_parent'];
             }
 
-            $response = $client->post("https://api.igdb.com/v4/alternative_names", [
-                'headers' => [
-                    'Content-Type' => 'application/json',
-                    'Accept' => 'application/json',
-                    'Client-ID' => 'sd5xdt5w2lkjr7ws92fxjdlicvb5u2',
-                    'Authorization' => $token
-                ],
-                'body' => 'fields *;where game = '.$idGame.' & comment = "French title";'
-            ]);
+            // ALTERNATIVE NAMES
 
-            $dataName = json_decode($response->getBody(), true);
+            $body = 'fields *;where game = '.$idGame.' & comment = "French title";';
+
+            $dataName = $IGDBService->getData('alternative_names', $body);
 
             if ($dataName !== []) {
                 $game->setName($dataName[0]['name']);
@@ -198,18 +181,9 @@ class GameController extends AbstractController
 
                 if (!$gameMode) {
 
+                    $body = 'fields name;where id = '.$gameModeId.';';
 
-                    $response = $client->post("https://api.igdb.com/v4/game_modes", [
-                        'headers' => [
-                            'Content-Type' => 'application/json',
-                            'Accept' => 'application/json',
-                            'Client-ID' => 'sd5xdt5w2lkjr7ws92fxjdlicvb5u2',
-                            'Authorization' => $token
-                        ],
-                        'body' => 'fields name;where id = '.$gameModeId.';'
-                    ]);
-
-                    $dataGameMode = json_decode($response->getBody(), true)[0];
+                    $dataGameMode = $IGDBService->getData('game_modes', $body);
 
                     $gameMode = new GameMode();
 
@@ -224,25 +198,16 @@ class GameController extends AbstractController
                 $game->addMode($gameMode);
             }
 
-            // GENRE
+            // GENRES
 
             foreach ($dataGame['genres'] as $genreId) {
                 $genre = $gameGenreRepository->findOneBy(['igdbId' => $genreId]);
 
                 if (!$genre) {
 
+                    $body = 'fields name;where id = '.$genreId.';';
 
-                    $response = $client->post("https://api.igdb.com/v4/genres", [
-                        'headers' => [
-                            'Content-Type' => 'application/json',
-                            'Accept' => 'application/json',
-                            'Client-ID' => 'sd5xdt5w2lkjr7ws92fxjdlicvb5u2',
-                            'Authorization' => $token
-                        ],
-                        'body' => 'fields name;where id = '.$genreId.';'
-                    ]);
-
-                    $dataGenre = json_decode($response->getBody(), true)[0];
+                    $dataGenre = $IGDBService->getData('genres', $body);
 
                     $genre = new GameGenre();
 
@@ -265,17 +230,9 @@ class GameController extends AbstractController
 
                 if (!$theme) {
 
-                    $response = $client->post("https://api.igdb.com/v4/themes", [
-                        'headers' => [
-                            'Content-Type' => 'application/json',
-                            'Accept' => 'application/json',
-                            'Client-ID' => 'sd5xdt5w2lkjr7ws92fxjdlicvb5u2',
-                            'Authorization' => $token
-                        ],
-                        'body' => 'fields name;where id = '.$themeId.';'
-                    ]);
+                    $body = 'fields name;where id = '.$themeId.';';
 
-                    $dataTheme = json_decode($response->getBody(), true)[0];
+                    $dataTheme = $IGDBService->getData('themes', $body);
 
                     $theme = new GameTheme();
 
@@ -293,17 +250,9 @@ class GameController extends AbstractController
 
             // RELEASE DATE
 
-            $response = $client->post("https://api.igdb.com/v4/release_dates", [
-                'headers' => [
-                    'Content-Type' => 'application/json',
-                    'Accept' => 'application/json',
-                    'Client-ID' => 'sd5xdt5w2lkjr7ws92fxjdlicvb5u2',
-                    'Authorization' => $token
-                ],
-                'body' => 'fields *;where game = '.$idGame.' & platform = '.$platformId.' & (status=6 | status = null) & (region=1 | region=8);'
-            ]);
+            $body = 'fields *;where game = '.$idGame.' & platform = '.$platformId.' & (status=6 | status = null) & (region=1 | region=8);';
 
-            $dataReleaseDate = json_decode($response->getBody(), true)[0];
+            $dataReleaseDate = $IGDBService->getData('release_dates', $body);
             $date = new DateTime();
             $date->setTimestamp($dataReleaseDate['date']);
 
@@ -311,17 +260,9 @@ class GameController extends AbstractController
 
             // PUBLISHER
 
-            $response = $client->post("https://api.igdb.com/v4/companies", [
-                'headers' => [
-                    'Content-Type' => 'application/json',
-                    'Accept' => 'application/json',
-                    'Client-ID' => 'sd5xdt5w2lkjr7ws92fxjdlicvb5u2',
-                    'Authorization' => $token
-                ],
-                'body' => 'fields name;where published = ['.$idGame.'];'
-            ]);
+            $body = 'fields name;where published = ['.$idGame.'];';
 
-            $dataPublisher = json_decode($response->getBody(), true);
+            $dataPublisher = $IGDBService->getData('companies', $body);
 
             foreach ($dataPublisher as $onePublisher) {
 
@@ -345,17 +286,9 @@ class GameController extends AbstractController
 
             // DEVELOPER
 
-            $response = $client->post("https://api.igdb.com/v4/companies", [
-                'headers' => [
-                    'Content-Type' => 'application/json',
-                    'Accept' => 'application/json',
-                    'Client-ID' => 'sd5xdt5w2lkjr7ws92fxjdlicvb5u2',
-                    'Authorization' => $token
-                ],
-                'body' => 'fields name;where developed = ['.$idGame.'];'
-            ]);
+            $body = 'fields name;where developed = ['.$idGame.'];';
 
-            $dataDeveloper = json_decode($response->getBody(), true);
+            $dataDeveloper = $IGDBService->getData('companies', $body);
 
             foreach ($dataDeveloper as $oneDeveloper) {
 
@@ -379,31 +312,14 @@ class GameController extends AbstractController
 
             // SERIES
 
-            $response = $client->post("https://api.igdb.com/v4/collections", [
-                'headers' => [
-                    'Content-Type' => 'application/json',
-                    'Accept' => 'application/json',
-                    'Client-ID' => 'sd5xdt5w2lkjr7ws92fxjdlicvb5u2',
-                    'Authorization' => $token
-                ],
-                'body' => 'fields id,name,games;where games = ['.$idGame.'];'
-            ]);
+            $body = 'fields id,name,games;where games = ['.$idGame.'];';
 
-            $dataSeries = json_decode($response->getBody(), true);
+            $dataSeries = $IGDBService->getData('collections', $body);
 
-            if (empty($dataSerie) && $idParent) {
+            if (empty($dataSeries) && $idParent) {
+                $body = 'fields id,name,games;where games = ['.$idParent.'];';
 
-                $response = $client->post("https://api.igdb.com/v4/collections", [
-                    'headers' => [
-                        'Content-Type' => 'application/json',
-                        'Accept' => 'application/json',
-                        'Client-ID' => 'sd5xdt5w2lkjr7ws92fxjdlicvb5u2',
-                        'Authorization' => $token
-                    ],
-                    'body' => 'fields id,name,games;where games = ['.$idParent.'];'
-                ]);
-
-                $dataSeries = json_decode($response->getBody(), true);
+                $dataSeries = $IGDBService->getData('collections', $body);
             }
 
             $saveSeries = null;
@@ -441,17 +357,9 @@ class GameController extends AbstractController
 
             if (!$platform) {
 
-                $response = $client->post("https://api.igdb.com/v4/platforms", [
-                    'headers' => [
-                        'Content-Type' => 'application/json',
-                        'Accept' => 'application/json',
-                        'Client-ID' => 'sd5xdt5w2lkjr7ws92fxjdlicvb5u2',
-                        'Authorization' => $token
-                    ],
-                    'body' => 'fields name;where id = '.$platformId.';'
-                ]);
+                $body = 'fields name;where id = '.$platformId.';';
 
-                $dataPlatform = json_decode($response->getBody(), true)[0];
+                $dataPlatform = $IGDBService->getData('platforms', $body);
 
                 $platform = new GamePlatform();
 
@@ -464,28 +372,16 @@ class GameController extends AbstractController
 
             // COVER
 
-            $response = $client->post("https://api.igdb.com/v4/covers", [
-                'headers' => [
-                    'Content-Type' => 'application/json',
-                    'Accept' => 'application/json',
-                    'Client-ID' => 'sd5xdt5w2lkjr7ws92fxjdlicvb5u2',
-                    'Authorization' => $token
-                ],
-                'body' => 'fields *;where game = '.$idGame.';'
-            ]);
+            $body = 'fields *;where game = '.$idGame.';';
 
-            $dataGameCover = json_decode($response->getBody(), true)[0];
+            $dataGameCover = $IGDBService->getData('covers', $body);
 
-            $lienImage = 'https:'.str_replace('/t_thumb/', '/t_cover_big/', $dataGameCover['url']);
+            $link = 'https:'.str_replace('/t_thumb/', '/t_cover_big/', $dataGameCover['url']);
 
-            $cover = imagecreatefromstring(file_get_contents($lienImage));
+            $destination = "/public/image/game/cover/".$game->getSlug().'.jpeg';
 
-            $projectDir = $kernel->getProjectDir();
-
-            $cheminImageDestination = "/public/image/game/cover/".$game->getSlug().'.jpeg';
-
-            if (imagejpeg($cover, $projectDir.$cheminImageDestination, 100)) {
-                $game->setCover($cheminImageDestination);
+            if ($fileService->addFile($link, $destination)) {
+                $game->setCover($destination);
             } else {
                 $game->setCover(null);
             }
@@ -507,6 +403,9 @@ class GameController extends AbstractController
     }
 
 
+    /**
+     * @throws Exception
+     */
     #[Route('/list', name: 'game_list', methods: ['POST'])]
     public function list(
         Request        $request,
@@ -516,6 +415,8 @@ class GameController extends AbstractController
     {
 
         $choice = $request->request->get('choice', 1);
+
+        /** @var ?string $text */
         $text = $request->request->get('text');
 
         $choice = intval($choice);
@@ -540,27 +441,23 @@ class GameController extends AbstractController
 
             $tooltip .= "<li>Date de sortie : ".$timeService->frenchFormatDateNoDay($game->getReleaseDate())."</li>";
 
-            if ($game->getGameTrackers() && $game->getGameTrackers()->getValues()){
+            if (count($game->getGameTrackers()) > 0 && $game->getGameTrackers()->getValues()) {
 
                 $gameTracker = $game->getGameTrackers()->getValues()[0];
 
-                if ($gameTracker->getStartDate()){
-                    $tooltip .= "<li>Commencer le : ".$timeService->frenchFormatDateNoDay($gameTracker->getStartDate())."</li>";
-                }
-
-                if ($gameTracker->getEndDate()){
+                if ($gameTracker->getEndDate()) {
                     $tooltip .= "<li>Fini le : ".$timeService->frenchFormatDateNoDay($gameTracker->getEndDate())."</li>";
                 }
 
-                if ($gameTracker->getEndTime()){
+                if ($gameTracker->getEndTime()) {
                     $tooltip .= "<li>Fini en : ".$timeService->convertirSecondes($gameTracker->getEndTime())."</li>";
                 }
 
-                if ($gameTracker->getCompleteDate()){
+                if ($gameTracker->getCompleteDate()) {
                     $tooltip .= "<li>Fini à 100% le : ".$timeService->frenchFormatDateNoDay($gameTracker->getCompleteDate())."</li>";
                 }
 
-                if ($gameTracker->getCompleteTime()){
+                if ($gameTracker->getCompleteTime()) {
                     $tooltip .= "<li>Fini à 100% en : ".$timeService->convertirSecondes($gameTracker->getCompleteTime())."</li>";
                 }
